Update cpu_optimizations tutorial.
I just tested colliding two rigid bodies each having 5000 Collision shape children and 5000 mesh instances with cube meshes, resulting in a total of 20 000 nodes in the tree. Godot does not seem to care the slightest... It takes a solid 10 seconds to spawn everything, but from there I have a solid 83% idle time according to the profiler (running a 4GHz pc that is).